pyproject.toml
[project]
name = "api-open-weather"
version = "0.1.0"
description = ""
authors = [
    {name = "brenno kenji",email = "brennokoj@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
